Add compute_timing_stats test and remove failing parse_frame
